Rust: Clean up the UnusedValue examples.

+++ rust/ql/src/queries/unusedentities/UnusedValueBad.rs
@@ -1,11 +1,11 @@
-fn get_average(values:&[i32]) -> i32 {
+fn get_average(values:&[i32]) -> f64 {
 	let mut sum = 0;
-	let mut average = 0; // BAD: unused value
+	let mut average = 0.0; // BAD: unused value
 
 	for v in values {
 		sum += v;
 	}
 
-        average = sum / values.len();
+	average = sum as f64 / values.len() as f64;
 	return average;
 }
